add updating si's details

+++ backend/internal/transport/http/v1/preservation/preservation.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Alexander272/mersi/backend/internal/models/response"
 	"github.com/Alexander272/mersi/backend/internal/services"
 	"github.com/Alexander272/mersi/backend/internal/transport/http/middleware"
+	"github.com/Alexander272/mersi/backend/internal/transport/http/utils"
 	"github.com/Alexander272/mersi/backend/pkg/error_bot"
 	"github.com/Alexander272/mersi/backend/pkg/logger"
 	"github.com/gin-gonic/gin"
@@ -87,6 +88,12 @@ func (h *Handler) create(c *gin.Context) {
 		return
 	}
 
+	actor := utils.GetActor(c)
+	if actor == nil {
+		return
+	}
+	dto.Actor = actor
+
 	if err := h.service.Create(c, dto); err != nil {
 		if errors.Is(err, models.ErrNotValid) {
 			response.NewErrorResponse(c, http.StatusBadRequest, err.Error(), "Отправлены некорректные данные")
@@ -114,6 +121,12 @@ func (h *Handler) update(c *gin.Context) {
 	}
 	dto.Id = id
 
+	actor := utils.GetActor(c)
+	if actor == nil {
+		return
+	}
+	dto.Actor = actor
+
 	if err := h.service.Update(c, dto); err != nil {
 		if errors.Is(err, models.ErrNotValid) {
 			response.NewErrorResponse(c, http.StatusBadRequest, err.Error(), "Отправлены некорректные данные")
@@ -135,6 +148,12 @@ func (h *Handler) delete(c *gin.Context) {
 		return
 	}
 	dto := &models.DeletePreservationDTO{Id: id}
+
+	actor := utils.GetActor(c)
+	if actor == nil {
+		return
+	}
+	dto.Actor = actor
 
 	if err := h.service.Delete(c, dto); err != nil {
 		response.NewErrorResponse(c, http.StatusInternalServerError, err.Error(), "Произошла ошибка: "+err.Error())
